+++ src/lightspeed_evaluation/core/llm/__init__.py
@@ -2,7 +2,7 @@
 
 from ..models import LLMConfig
 from .deepeval import DeepEvalLLMManager
-from .manager import LLMError, LLMManager
+from .manager import LLMError, LLMManager, validate_gemini_env, validate_openai_env
 from .ragas import RagasLLMManager
 
 __all__ = [
@@ -11,4 +11,6 @@ __all__ = [
     "LLMManager",
     "DeepEvalLLMManager",
     "RagasLLMManager",
+    "validate_openai_env",
+    "validate_gemini_env",
 ]
